requirements.txt
Flask==3.0.3
flask-sock==0.7.0
Flask-Cors==5.0.0
uvicorn==0.30.3
gunicorn==21.2.0
numpy==2.1.3
opencv-python==4.11.0.86
Pillow==11.0.0
ultralytics==8.3.160
dashscope==1.24.6
PyYAML==6.0.2
pydantic==2.9.2
pydantic-settings==2.6.1
ffmpeg-python==0.2.0
requests==2.32.3
loguru==0.7.2
python-dotenv==1.0.1
websocket-client==1.8.0
